@@ -1,21 +1,29 @@
 {
+  "name": "simulator-ui",
+  "version": "1.0.1",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "serve": "ng serve --proxy-config proxy.conf.json",
+    "lite": "lite-server",
+    "tsc": "tsc -p src/",
+    "tsc:w": "tsc -w",
+    "build": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
   "private": true,
   "dependencies": {
-    "@angular/common": "^2.4.6",
-    "@angular/compiler": "^2.4.6",
-    "@angular/core": "^2.4.6",
-    "@angular/forms": "^2.4.6",
-    "@angular/http": "^2.4.6",
-    "@angular/platform-browser": "^2.4.6",
-    "@angular/platform-browser-dynamic": "^2.4.6",
-    "@angular/router": "^3.3.1",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^2.0.0",
-    "@ngrx/router-store": "^1.2.5",
-    "@ngrx/store": "^2.2.1",
-    "@ngrx/store-devtools": "^3.2.3",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "ace": "^1.3.0",
-    "angular-in-memory-web-api": "~0.1.5",
     "bootstrap": "^3.3.7",
     "brace": "^0.9.1",
     "core-js": "^2.4.1",
@@ -25,47 +33,39 @@
     "lodash": "^4.17.4",
     "moment": "2.17.1",
     "reflect-metadata": "^0.1.8",
-    "rxjs": "^5.1.0",
+    "rxjs": "^5.5.2",
     "sockjs-client": "^1.1.2",
     "stompjs": "2.3.3",
     "systemjs": "0.19.39",
     "underscore": "1.8.3",
     "url": "^0.11.0",
-    "zone.js": "^0.7.6"
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^2.3.1",
+    "@angular/cli": "1.5.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/language-service": "^5.0.0",
+    "@types/jasmine": "~2.5.53",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
     "@types/ace": "0.0.32",
-    "@types/jasmine": "2.5.38",
     "@types/jquery": "^2.0.40",
     "@types/lodash": "^4.14.58",
-    "@types/node": "^7.0.11",
     "@types/sockjs": "^0.3.30",
     "@types/sockjs-client": "^1.0.30",
     "@types/stompjs": "^2.3.0",
-    "angular-cli": "1.0.0-beta.28.3",
-    "codelyzer": "~2.0.0-beta.1",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "~4.0.13",
-    "ts-node": "1.2.1",
-    "tslint": "^4.3.0",
-    "typescript": "2.2.1"
-  },
-  "scripts": {
-    "start": "ng serve",
-    "serve": "ng serve --proxy-config proxy.conf.json",
-    "lite": "lite-server",
-    "tsc": "tsc -p src/",
-    "tsc:w": "tsc -w",
-    "build": "ng build --prod",
-    "test": "ng test",
-    "pree2e": "webdriver-manager update --standalone false --gecko false",
-    "e2e": "protractor"
+    "codelyzer": "~3.2.0",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.1.2",
+    "ts-node": "~3.2.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.4.2"
   }
 }
