@@ -19,6 +19,8 @@ package com.consol.citrus.simulator.sample.starter;
 import com.consol.citrus.simulator.model.ScenarioParameter;
 import com.consol.citrus.simulator.model.ScenarioParameterBuilder;
 import com.consol.citrus.simulator.scenario.*;
+import com.consol.citrus.simulator.service.TemplateService;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -28,6 +30,9 @@ import java.util.List;
  */
 @Starter("HelloStarter")
 public class HelloStarter extends AbstractScenarioStarter {
+
+    @Autowired
+    private TemplateService templateService;
 
     @Override
     public void run(ScenarioRunner scenario) {
@@ -54,7 +59,7 @@ public class HelloStarter extends AbstractScenarioStarter {
                 .label("Payload")
                 .required()
                 .textarea()
-                .value(getXmlMessageTemplate("Hello"))
+                .value(templateService.getXmlMessageTemplate("Hello"))
                 .build());
 
         return scenarioParameters;
