@@ -16,7 +16,9 @@
 
 package com.consol.citrus.simulator.sample.scenario;
 
-import com.consol.citrus.simulator.scenario.*;
+import com.consol.citrus.simulator.scenario.AbstractSimulatorScenario;
+import com.consol.citrus.simulator.scenario.Scenario;
+import com.consol.citrus.simulator.scenario.ScenarioRunner;
 import org.springframework.http.HttpStatus;
 
 /**
@@ -26,14 +28,16 @@ import org.springframework.http.HttpStatus;
 public class DefaultScenario extends AbstractSimulatorScenario {
 
     @Override
-    public void run(ScenarioDesigner scenario) {
+    public void run(ScenarioRunner scenario) {
         scenario
-            .receive();
+            .http()
+            .receive((builder -> builder.post()));
 
         scenario
             .http()
-            .send()
-            .response(HttpStatus.OK)
-            .payload("<DefaultResponse>This is a default response!</DefaultResponse>");
+            .send((builder -> builder
+                    .response(HttpStatus.OK)
+                    .payload("<DefaultResponse>This is a default response!</DefaultResponse>"))
+            );
     }
 }
