@@ -16,7 +16,9 @@
 
 package com.consol.citrus.simulator.sample.scenario;
 
-import com.consol.citrus.simulator.scenario.*;
+import com.consol.citrus.simulator.scenario.AbstractSimulatorScenario;
+import com.consol.citrus.simulator.scenario.Scenario;
+import com.consol.citrus.simulator.scenario.ScenarioRunner;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -29,21 +31,23 @@ import org.springframework.web.bind.annotation.RequestMethod;
 public class GoodByeScenario extends AbstractSimulatorScenario {
 
     @Override
-    public void run(ScenarioDesigner scenario) {
+    public void run(ScenarioRunner scenario) {
         scenario
             .http()
-            .receive()
-            .post()
-            .payload("<GoodBye xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Say GoodBye!" +
-                     "</GoodBye>");
+            .receive((builder -> builder
+                    .post()
+                    .payload("<GoodBye xmlns=\"http://citrusframework.org/schemas/hello\">" +
+                            "Say GoodBye!" +
+                            "</GoodBye>"))
+            );
 
         scenario
             .http()
-            .send()
-            .response(HttpStatus.OK)
-            .payload("<GoodByeResponse xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Bye bye!" +
-                     "</GoodByeResponse>");
+            .send((builder -> builder
+                    .response(HttpStatus.OK)
+                    .payload("<GoodByeResponse xmlns=\"http://citrusframework.org/schemas/hello\">" +
+                            "Bye bye!" +
+                            "</GoodByeResponse>"))
+            );
     }
 }
