@@ -19,6 +19,8 @@ package com.consol.citrus.simulator.sample.starter;
 import com.consol.citrus.simulator.model.ScenarioParameter;
 import com.consol.citrus.simulator.model.ScenarioParameterBuilder;
 import com.consol.citrus.simulator.scenario.*;
+import com.consol.citrus.simulator.service.TemplateService;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.*;
 
@@ -27,6 +29,9 @@ import java.util.*;
  */
 @Starter("GoodByeStarter")
 public class GoodByeStarter extends AbstractScenarioStarter {
+
+    @Autowired
+    private TemplateService templateService;
 
     @Override
     public void run(ScenarioDesigner scenario) {
@@ -43,7 +48,7 @@ public class GoodByeStarter extends AbstractScenarioStarter {
                 .label("Payload")
                 .required()
                 .textarea()
-                .value(getXmlMessageTemplate("Goodbye"))
+                .value(templateService.getXmlMessageTemplate("Goodbye"))
                 .build());
 
         return scenarioParameter;
