@@ -16,7 +16,9 @@
 
 package com.consol.citrus.simulator.sample.scenario;
 
-import com.consol.citrus.simulator.scenario.*;
+import com.consol.citrus.simulator.scenario.AbstractSimulatorScenario;
+import com.consol.citrus.simulator.scenario.Scenario;
+import com.consol.citrus.simulator.scenario.ScenarioDesigner;
 
 /**
  * @author Christoph Deppisch
@@ -30,15 +32,15 @@ public class GoodByeScenario extends AbstractSimulatorScenario {
             .soap()
             .receive()
             .payload("<GoodBye xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Say GoodBye!" +
-                     "</GoodBye>")
+                    "Say GoodBye!" +
+                    "</GoodBye>")
             .soapAction("GoodBye");
 
         scenario
             .soap()
             .send()
             .payload("<GoodByeResponse xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Bye bye!" +
-                     "</GoodByeResponse>");
+                    "Bye bye!" +
+                    "</GoodByeResponse>");
     }
 }
