@@ -19,7 +19,9 @@ package com.consol.citrus.simulator.sample.scenario;
 import com.consol.citrus.context.TestContext;
 import com.consol.citrus.endpoint.adapter.mapping.XPathPayloadMappingKeyExtractor;
 import com.consol.citrus.message.Message;
-import com.consol.citrus.simulator.scenario.*;
+import com.consol.citrus.simulator.scenario.AbstractSimulatorScenario;
+import com.consol.citrus.simulator.scenario.Scenario;
+import com.consol.citrus.simulator.scenario.ScenarioDesigner;
 
 /**
  * @author Christoph Deppisch
@@ -30,14 +32,14 @@ public class GoodNightScenario extends AbstractSimulatorScenario {
     @Override
     public void run(ScenarioDesigner scenario) {
         scenario.correlation().start()
-            .withHandler(this);
+                .withHandler(this);
 
         scenario
             .soap()
             .receive()
             .payload("<GoodNight xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Go to sleep!" +
-                     "</GoodNight>")
+                    "Go to sleep!" +
+                    "</GoodNight>")
             .soapAction("GoodNight");
 
         scenario
@@ -50,16 +52,16 @@ public class GoodNightScenario extends AbstractSimulatorScenario {
             .soap()
             .receive()
             .payload("<GoodNight xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Go to sleep!" +
-                     "</GoodNight>")
+                    "Go to sleep!" +
+                    "</GoodNight>")
             .soapAction("GoodNight");
 
         scenario
             .soap()
             .send()
             .payload("<GoodNightResponse xmlns=\"http://citrusframework.org/schemas/hello\">" +
-                        "Good Night!" +
-                     "</GoodNightResponse>");
+                    "Good Night!" +
+                    "</GoodNightResponse>");
     }
 
     @Override
