@@ -23,7 +23,7 @@ import com.consol.citrus.simulator.correlation.CorrelationHandler;
 import com.consol.citrus.simulator.correlation.CorrelationHandlerRegistry;
 import com.consol.citrus.simulator.exception.SimulatorException;
 import com.consol.citrus.simulator.scenario.SimulatorScenario;
-import com.consol.citrus.simulator.service.ScenarioService;
+import com.consol.citrus.simulator.service.ScenarioExecutionService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,15 +32,17 @@ import org.springframework.context.ApplicationContextAware;
 import org.springframework.util.StringUtils;
 
 import java.util.Collections;
-import java.util.concurrent.*;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 /**
  * @author Christoph Deppisch
  */
 public class SimulatorEndpointAdapter extends RequestDispatchingEndpointAdapter implements ApplicationContextAware {
 
-    /** Logger */
-    private static Logger LOG = LoggerFactory.getLogger(SimulatorEndpointAdapter.class);
+    private static final Logger LOG = LoggerFactory.getLogger(SimulatorEndpointAdapter.class);
 
     @Autowired
     private CorrelationHandlerRegistry handlerRegistry;
@@ -49,12 +51,16 @@ public class SimulatorEndpointAdapter extends RequestDispatchingEndpointAdapter 
     private SimulatorConfigurationProperties configuration;
 
     @Autowired
-    private ScenarioService scenarioService;
+    private ScenarioExecutionService scenarioExecutionService;
 
-    /** Spring application context */
+    /**
+     * Spring application context
+     */
     private ApplicationContext applicationContext;
 
-    /** When adapter is asynchronous response handling is skipped */
+    /**
+     * When adapter is asynchronous response handling is skipped
+     */
     private boolean handleResponse = true;
 
     @Override
@@ -73,7 +79,10 @@ public class SimulatorEndpointAdapter extends RequestDispatchingEndpointAdapter 
             } catch (TimeoutException e) {
                 LOG.warn(String.format("No response for scenario '%s'", handler.getScenarioEndpoint().getName()));
                 return null;
-            } catch (InterruptedException | ExecutionException e) {
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                throw new SimulatorException(e);
+            } catch (ExecutionException e) {
                 throw new SimulatorException(e);
             }
         } else {
@@ -90,15 +99,15 @@ public class SimulatorEndpointAdapter extends RequestDispatchingEndpointAdapter 
             scenario = applicationContext.getBean(scenarioName, SimulatorScenario.class);
         } else {
             scenarioName = configuration.getDefaultScenario();
-            LOG.info(String.format("Unable to find scenario for mapping '%s' - " +
-                    "using default scenario '%s'", mappingName, scenarioName));
+            LOG.info("Unable to find scenario for mapping '{}' - " +
+                    "using default scenario '{}'", mappingName, scenarioName);
             getTestContext().getGlobalVariables().put("scenario_mapping_name", mappingName);
             scenario = applicationContext.getBean(scenarioName, SimulatorScenario.class);
         }
 
         scenario.getScenarioEndpoint().setName(scenarioName);
         scenario.getScenarioEndpoint().add(request, responseFuture);
-        scenarioService.run(scenario, scenarioName, Collections.EMPTY_LIST);
+        scenarioExecutionService.run(scenario, scenarioName, Collections.emptyList());
 
         try {
             if (handleResponse) {
@@ -109,7 +118,10 @@ public class SimulatorEndpointAdapter extends RequestDispatchingEndpointAdapter 
         } catch (TimeoutException e) {
             LOG.warn(String.format("No response for scenario '%s'", scenarioName));
             return null;
-        } catch (InterruptedException | ExecutionException e) {
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new SimulatorException(e);
+        } catch (ExecutionException e) {
             throw new SimulatorException(e);
         }
     }
