@@ -18,8 +18,14 @@ package com.consol.citrus.simulator.scenario.mapper;
 
 import com.consol.citrus.endpoint.adapter.mapping.JsonPayloadMappingKeyExtractor;
 import com.consol.citrus.message.Message;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.util.StringUtils;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Predicate;
 
 /**
  * Mapper that can be used for examining incoming messages and based on the content of the message returns the name
@@ -29,11 +35,13 @@ import java.util.*;
  * message.
  */
 public class ContentBasedJsonPathScenarioMapper implements ScenarioMapper {
-
+    private static final Logger LOG = LoggerFactory.getLogger(ContentBasedJsonPathScenarioMapper.class);
     private final List<JsonPayloadMappingKeyExtractor> keyExtractors = new ArrayList<>();
+    private Predicate<String> mappingKeyFilter = StringUtils::hasLength;
 
     /**
      * Fluent API builder method adds new json path expression.
+     *
      * @param jsonPathExpression
      * @return
      */
@@ -43,7 +51,20 @@ public class ContentBasedJsonPathScenarioMapper implements ScenarioMapper {
     }
 
     /**
+     * Fluent API builder method adds scenarioNameFilter. By adding a filter you can limit the mapped keys to a
+     * particular set of scenario names or to a particular pattern.
+     *
+     * @param mappingKeyFilter for filtering out or skipping mapped keys
+     * @return
+     */
+    public ContentBasedJsonPathScenarioMapper addMappingKeyFilter(Predicate<String> mappingKeyFilter) {
+        this.mappingKeyFilter = mappingKeyFilter;
+        return this;
+    }
+
+    /**
      * Create proper json path mapping key extractor from given expression.
+     *
      * @param jsonPathExpression
      * @return
      */
@@ -55,22 +76,33 @@ public class ContentBasedJsonPathScenarioMapper implements ScenarioMapper {
 
     @Override
     public String extractMappingKey(Message request) {
-        Optional<String> v = keyExtractors.parallelStream()
+        Optional<String> v = keyExtractors.stream()
                 .map(keyExtractor -> lookupScenarioName(request, keyExtractor))
                 .filter(Optional::isPresent)
                 .map(Optional::get)
-                .findAny();
+                .filter(mappingKeyFilter)
+                .findFirst();
         return v.orElse(null);
     }
 
     /**
      * Look up scenario name for given request.
+     *
      * @param request
      * @param keyExtractor
      * @return
      */
     private Optional<String> lookupScenarioName(Message request, JsonPayloadMappingKeyExtractor keyExtractor) {
-        return Optional.ofNullable(keyExtractor.getMappingKey(request));
+        try {
+            final String mappingKey = keyExtractor.getMappingKey(request);
+            LOG.debug("Scenario-name lookup returned: {}", mappingKey);
+            return Optional.of(mappingKey);
+        } catch (RuntimeException e) {
+            LOG.trace("Scenario-name lookup failed", e);
+        }
+
+        LOG.debug("Scenario-name lookup returned: <no-match>");
+        return Optional.empty();
     }
 }
 
