@@ -4,13 +4,16 @@ import com.consol.citrus.http.message.HttpMessage;
 import com.consol.citrus.simulator.config.SimulatorConfigurationProperties;
 import com.consol.citrus.simulator.scenario.AbstractSimulatorScenario;
 import com.consol.citrus.simulator.scenario.Scenario;
-import org.mockito.*;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.springframework.http.HttpMethod;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 
 /**
  * @author Christoph Deppisch
@@ -32,66 +35,45 @@ public class HttpRequestAnnotationScenarioMapperTest {
 
     @Test
     public void testGetMappingKey() {
-        scenarioMapper.getScenarios().add(new FooListScenario());
+        scenarioMapper.getScenarios().add(new IssueScenario());
         scenarioMapper.getScenarios().add(new FooScenario());
-        scenarioMapper.getScenarios().add(new FooDetailScenario());
-        scenarioMapper.getScenarios().add(new BarListScenario());
-        scenarioMapper.getScenarios().add(new BarScenario());
-        scenarioMapper.getScenarios().add(new BarDetailScenario());
+        scenarioMapper.getScenarios().add(new GetFooScenario());
+        scenarioMapper.getScenarios().add(new PutFooScenario());
+        scenarioMapper.getScenarios().add(new OtherScenario());
 
-        HttpMessage request = new HttpMessage();
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "default");
-
-        request.path("/issues");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "default");
-
-        request.path("/issues/foos");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "fooListScenario");
-
-        request.path("/issues/bars");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "barListScenario");
-
-        request.path("/issues/foo/1");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "fooScenario");
-
-        request.path("/issues/bar/1");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "barScenario");
-
-        request.path("/issues/foo/detail");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "fooDetailScenario");
-
-        request.path("/issues/bar/detail");
-        Assert.assertEquals(scenarioMapper.getMappingKey(request), "barDetailScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/foo")), "FooScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/foo").method(HttpMethod.GET)), "GetFooScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/foo").method(HttpMethod.PUT)), "PutFooScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/other")), "OtherScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/bar").method(HttpMethod.GET)), "IssueScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/bar").method(HttpMethod.DELETE)), "IssueScenario");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/bar").method(HttpMethod.PUT)), "default");
+        Assert.assertEquals(scenarioMapper.getMappingKey(new HttpMessage().path("/issues/bar")), "default");
+        Assert.assertEquals(scenarioMapper.getMappingKey(null), "default");
     }
 
-    @Scenario("fooListScenario")
-    @RequestMapping("/issues/foos")
-    private class FooListScenario extends AbstractSimulatorScenario {
-    }
-
-    @Scenario("fooScenario")
-    @RequestMapping("/issues/foo/{id}")
+    @Scenario("FooScenario")
+    @RequestMapping(value = "/issues/foo", method = RequestMethod.POST)
     private class FooScenario extends AbstractSimulatorScenario {
     }
 
-    @Scenario("fooDetailScenario")
-    @RequestMapping("/issues/foo/detail")
-    private class FooDetailScenario extends AbstractSimulatorScenario {
+    @Scenario("GetFooScenario")
+    @RequestMapping(value = "/issues/foo", method = RequestMethod.GET)
+    private class GetFooScenario extends AbstractSimulatorScenario {
     }
 
-    @Scenario("barListScenario")
-    @RequestMapping("/issues/bars")
-    private class BarListScenario extends AbstractSimulatorScenario {
+    @Scenario("PutFooScenario")
+    @RequestMapping(value = "/issues/foo", method = RequestMethod.PUT)
+    private class PutFooScenario extends AbstractSimulatorScenario {
     }
 
-    @Scenario("barScenario")
-    @RequestMapping("/issues/bar/{id}")
-    private class BarScenario extends AbstractSimulatorScenario {
+    @Scenario("IssueScenario")
+    @RequestMapping(value = "/issues/{name}", method = { RequestMethod.GET, RequestMethod.DELETE })
+    private class IssueScenario extends AbstractSimulatorScenario {
     }
 
-    @Scenario("barDetailScenario")
-    @RequestMapping("/issues/bar/detail")
-    private class BarDetailScenario extends AbstractSimulatorScenario {
+    @Scenario("OtherScenario")
+    @RequestMapping("/issues/other")
+    private class OtherScenario extends AbstractSimulatorScenario {
     }
-
 }
