@@ -1,38 +1,28 @@
 // Karma configuration file, see link for more information
-// https://karma-runner.github.io/0.13/config/configuration-file.html
+// https://karma-runner.github.io/1.0/config/configuration-file.html
 
 module.exports = function (config) {
     config.set({
         basePath: '',
-        frameworks: ['jasmine', 'angular-cli'],
+        frameworks: ['jasmine', '@angular/cli'],
         plugins: [
             require('karma-jasmine'),
             require('karma-chrome-launcher'),
-            require('karma-remap-istanbul'),
-            require('angular-cli/plugins/karma')
+            require('karma-jasmine-html-reporter'),
+            require('karma-coverage-istanbul-reporter'),
+            require('@angular/cli/plugins/karma')
         ],
-        files: [
-            { pattern: './src/test.ts', watched: false }
-        ],
-        preprocessors: {
-            './src/test.ts': ['angular-cli']
+        client:{
+            clearContext: false // leave Jasmine Spec Runner output visible in browser
         },
-        mime: {
-            'text/x-typescript': ['ts','tsx']
-        },
-        remapIstanbulReporter: {
-            reports: {
-                html: 'coverage',
-                lcovonly: './coverage/coverage.lcov'
-            }
+        coverageIstanbulReporter: {
+            reports: [ 'html', 'lcovonly' ],
+            fixWebpackSourcePaths: true
         },
         angularCli: {
-            config: './angular-cli.json',
             environment: 'dev'
         },
-        reporters: config.angularCli && config.angularCli.codeCoverage
-            ? ['progress', 'karma-remap-istanbul']
-            : ['progress'],
+        reporters: ['progress', 'kjhtml'],
         port: 9876,
         colors: true,
         logLevel: config.LOG_INFO,
