@@ -14,7 +14,8 @@ export class ScenarioExecution {
         public status: string,
         public errorMessage: string,
         public scenarioParameters: Array<ScenarioParameter>,
-        public scenarioActions: Array<ScenarioAction>) {
+        public scenarioActions: Array<ScenarioAction>,
+        public scenarioMessages: Array<Message>) {
     }
 }
 
@@ -64,6 +65,16 @@ export class Message {
         public messageId: number,
         public direction: string,
         public date: number,
-        public payload: string) {
+        public scenarioExecutionId: number,
+        public scenarioName: string,
+        public payload: string,
+        public headers: Array<MessageHeader>) {
+    }
+}
+
+export class MessageHeader {
+    constructor(
+        public name: string,
+        public value: string) {
     }
 }
